pyproject.toml
[project]
name = "agent-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyjwt[crypto]",
    "requests",
    "python-dotenv",
    "pyyaml",
    "dashscope",
    "pygithub"
]
